+++ lib/ts/recipe/session/recipe.ts
@@ -13,7 +13,7 @@
  * under the License.
  */
 import RecipeModule from "../recipeModule";
-import SuperTokensWebsite from "supertokens-website";
+import SuperTokensWebsite, { ClaimValidationError, SessionClaimValidator } from "supertokens-website";
 import { InputType, UserInput } from "./types";
 import { checkForSSRErrorAndAppendIfNeeded, isTest } from "../../utils";
 import { CreateRecipeFunction, NormalisedAppInfo } from "../../types";
@@ -89,6 +89,24 @@ export default class Recipe extends RecipeModule<unknown, any> {
     attemptRefreshingSession = async (): Promise<boolean> => {
         return SuperTokensWebsite.attemptRefreshingSession();
     };
+
+    validateClaims = (input: {
+        overrideGlobalClaimValidators?: (
+            globalClaimValidators: SessionClaimValidator[],
+            userContext: any
+        ) => SessionClaimValidator[];
+        userContext: any;
+    }): Promise<ClaimValidationError[]> | ClaimValidationError[] => {
+        return SuperTokensWebsite.validateClaims(input.overrideGlobalClaimValidators, input.userContext);
+    };
+
+    // The strange typing is to avoid adding a dependency to axios
+    getInvalidClaimsFromResponse(input: {
+        response: { data: any } | Response;
+        userContext: any;
+    }): Promise<ClaimValidationError[]> {
+        return SuperTokensWebsite.getInvalidClaimsFromResponse(input);
+    }
 
     static addAxiosInterceptors(axiosInstance: any, userContext: any): void {
         return SuperTokensWebsite.addAxiosInterceptors(axiosInstance, userContext);